+++ nodalia-cards.manifest.js
@@ -1,8 +1,8 @@
 export default {
-  "pkgVersion": "1.1.0-alpha.1",
-  "contentSha256_12": "5b324269534a",
+  "pkgVersion": "1.1.0-alpha.2",
+  "contentSha256_12": "b5099cb0929d",
   "file": "nodalia-cards.bundle.js"
 };
-export const pkgVersion = "1.1.0-alpha.1";
-export const contentSha256_12 = "5b324269534a";
+export const pkgVersion = "1.1.0-alpha.2";
+export const contentSha256_12 = "b5099cb0929d";
 export const file = "nodalia-cards.bundle.js";